Hello there,
Firstly, I wanted to say say huge thank you for the package that you have provided. Very great piece of software that helps with spreadsheets.

Secondly, I noticed that security solution (SNYK) I use for my application pointed out few issues, which are only related to packages used by ClosedXML. I have pulled the branch and updated everything just to make sure that will be resolved and it occured that this simple procedure fixed everything.

Please check the updated packages and if it is okay merge it, so I could keep using the ClosedXML as nuget package instead of exporting it ito bunch of dll files.

Kindest regards,
Ben