#2066 Embedd an ultimate fallback font directly into the assembly. Due to fragmentation, there is no reasonable font I could rely on all environments.
Although it takes up 100k, it is a significant life improvement for users of the library. Since Carlito is metric compatible with Calibri and Calibri is a default font, the auto sizing will work reasonable accurately and columns will be adjusted to expected width.

The Carlito font is under SIL Open License, so I can strip all unnecessary information and embedd it (see SIL FAQ 1.10 and 1.11).