An initial work on a dependency tree for formulas. The tree can mark formulas as dirty, if some precedent (direct/transitive) cell is changed.
The chosen data structure is an R-tree. R-tree seems to be the only reasonable structure (other than the variant R*-tree) for orthogonal range searching for overlapping ranges, most other are point vs range. Others are mostly "variants" of the main algorithm that is intended for other things (e.g. kd-tree).