Separate a function that modifies the style of the XLCell when it has its value set. The separation will help to reuse same logic for InsertData without instatntiating cells.
Start a new paradigm for modifying styles. Originally, the only way to modify the style was to instantiate a new XLStyle (through InnerStyle) and then call setter. The new way uses a modification of a StyleValue and then sets the modified StyleValue. It only allocates something, if a new style is created, because otherwise the {Style,Alignment,Font...}Value is taken from repository.