Save pivot cache records to the output file.
This will ensure, that what was loaded will be saved to the output file. Originally, the cache records were lost and thus any change to the pivot table was impossible for workbooks with changed source (deleted workbook, removed column...). Thanks to round trip load/save, that is no longer the case.

The records for new pivot tables need work, they duplicate values instead of refering to the shared items in pivot cache definition, but Excel can read the records anyway, so it is a significant is progress.