Update language level and use polyfill for missing functions.
The Polyfill library is a development only dependency that has *.cs sources in the .nuget directory. The *.cs files are conditionaly compiled along the project depending on the target and thus more modern version use less polyfills.

The main benefit is ability to use struct records (that require IsExternalInit attribute for init properties), though generally it just allows a modern development even for netstandard2.0.

The polyfills are internal and thus don't pollute applications consuming ClosedXML.