Recognize day/months in date to number conversion even without leading zeros.
Many cultures return only formats with leading zeros from DateTimeFormat.GetAllDateTimePatterns(), thus the attempt to recognize the date is not as fuzzy as it should be, e.g. 2026/4/3 might not be recognized, because pattern si only yyyy/dd/MM. The fuzzy parser replaces MM/dd with M/d to be more inclusive.