Width of a table (col E) in sheet4 is now correct according to adjust to content.
The original file was incorrect, because the column 5 has been inserted to a
table, but didn't have a column set a value to a cell. Because there was no
value in cell, adjust to content didn't do its work properly. With structured
references, teh adjust to content called GetFormattedString which called Value
which called recalculatethat touched touch field names and thus are rescaned...

Whole XLTable will have to be reworked anyway, because of improper
materialization and other stuff.