Load all PivotTableDefinition fields into the XLPivotTable.
XLPivotTable didn't have properties for all attributes of
PivotTableDefinition. Because of it, it wasn't able to be a "dump pipe"
and just write what it read.

XLPivotTable now contains properties for all PTD attributes and reads
them. It will facilitate full representation, so we can just write the
output without patching original XML.