Add a skeleton of a pivot table stream writer.
This is a part of pivot table rework. There are newly filled structures
(without any checking) and the goal is to be able to "pipe" pivot table
through, as long as it's not modified and still create valid pivot table.

This adds ability to create a empty pivot table part with empty fields, if
writer is switched in XLWorkbook_Save. Currently, switch is not done and
original writer is still used to keep code dormant while it is being worked on.

Currently, the pivot table is written with PivotTableDefinitionPartWriter
that uses OpenXML SDK structures. We want to move away from OpenXML
SDK to a more streaming solution (plus more performanct).