Add loading and writing pivot tabel styles to Writer2.
Old code has a difficulty reading style because it needs to translate
from file structures to data structures. The new structures are more
in-line with a file, leading to easier loading and saving.

This adds ability to read and write formats, so it can work with
more pivot tables in various files. The code is in Writer2 that is not
used in rest of codebase yet.

We have no special object for dxf, so use XLStyle. That makes me unhappy,
but a lot of things make me unhappy.