Add pivot table conditional formats load/save.
The pivot tables weren't able to represent conditional formatting,
this adds structures and load and save code. The saving code is in
Writer2, which is not yet enabled.

Even when Writer2 is used, it doesn't work, because it needs implicit
relationship with worksheet part conditional formatting elements.
There must be CF in the PT sheet with pivot flag and same priority.
Without it, Excel tries to repair file.