Unify page fields on pivot table.
There are two types of objects: one that contains data
for the page fields and other that provides API to the user.

Originally, they weren't unified and had different data.

Now the page fields resemble data fields and the API objects
are created as needed. The problem with inconsistency is
gone, because everything is derived from source of truth.