Fix MaintainPivotTableLabelsOrder test.
The test was failing because of non-unified page fields
(fixed in previsou commit) and double loading of data field.
The data field on an axis was already set (indicated by set
DataPosition value) and when axes encountered 'data' field
and was trying to add it to the axes (XLPivotTableAxis.AddField),
it failed, because the field was already there.