API to add/modify grand total styles
Styling objects are basically only proxies that are used to modify the
XLPivotFormat collection (that is basically a dictionary of area - dx
style) of the underlaying data structure of pivot tables.

Also implement IXStylized in XLPivotFormat. That is necessary, because only
then can XLPivotFormat be a container that is notified about changes in a
style.

XLPivotFormat woudn't be notified about style changes
and would thus keep stale style and the output file wouldn't
contain desired style without the IXLStylized.