Add style API for pivot field headers
The API was removed during refactoring and thus it wasn't possible to
change style of a pivot fields header.

The headers are displayed either individually for each field or together.
Which one is used depends on the values in pivot fields, but they affest
shared settings and thus aren't individual.

This uses classic API object with areas, same as all other pivot styling options.