Move hyperlink from slice to XLHyperlinks
The key benefit of this move is that the source of the truth moves to
one place: XLHyperlinks. Originally, there were three places with a
location of a hyperlink:

* XLHyperlink.Cell
* XLHyperlinks dictionary of ranges
* MiscSlice

The hyperlink is not in really a part of slices. Slices are for things
inherently attached to cells. Hyperlink is not inherently part of the
cell. It is a property of an area, just like merged ranges, table or
pivot table. It is an additional property of a range, not part of
a cell.