Hyperlinks reacts on structual changes
This is the first proper implementation of ISheetListener. Currently,
we rely on a IXLRange repository that moves ranges around on structural
changes.

It has several drawbacks:

* XLWorksheet is a god class that handles all shifting stuff.
* RangeRepository keeps track of all ranges. Any time a range is
  created, it is added to repository and when any structural change
  happens, every tracked range is updated. That can get really
  expensive and isn't transparent to the user.
* The whole thing is not modular and rather of complicated.
* Doesn't support insertion of areas, only insertion of rows and
  columns.

Instead, I want each component to deal with it on it's own (e.g.
hyperlink, merged range, tables, sparklines...). That will make it more
modular and hopefully simpler to reason about.

In most cases, it will be simple shift, like in the hyperlink case, but
in many other cases there could be a more complex logic (e.g. shifted
merged range, shift headers/footers of a table...). The complex logic
should be encapsulated in the component itself, not XLWorksheet.

In the long run, I hope to remove need to track and update
size/position of XLRange, just like I did for XLCell.