Split tally calculation
Tally calculation was done in one function and it was rather hard to
read, requiring a delegate with many parameters and functions overload.

This splits tally calculation into two parts:
* function converting arguments into scalar values
* actual algorithm that aggregates values into a single value (e.g. sum
  or average).

The ITally implementations (TallyNumbers and TallyAll) will be reusable
across many functions (though Excel is inconsistent, there is only
limited number of ways it counts or doesn't count some values).

Because ITally is an interface, I will also be able to provide custom
substotal ITally that filters out cells with SUBTOTAL function or that
are hidden. Thus I will be able to reuse functions in subtotal (e.g.
AVERAGE) and only supply different ITally implementation.