Update DocumentFormat.OpenXml due to vulnerability
The DocumentFormat.OpenXml depends on System.IO.Packaging. We originally
(when we depended on OpenXML SDK 2.x) used explicit dependency to solve
a problem with race condition.

The OpenXML SDK  3* depends on System.IO.Packaging 8.* anyway, so there is
no need to explicitely depend on it anymore and it can be removed.

The System.IO.Packaging 8.0.0 and earlier have two vulnerabilities:
* GHSA-qj66-m88j-hmgj
* GHSA-f32c-w444-8ppv

The 3.1.1 version depends on System.IO.Packaging 8.0.1 and thus we no
longer need explicitly transitive and solve security issue.