Update rich text when it changes
We have mutable API for rich text string, but the implementation must be immutable. Immutability is required by shared string table, that stores whole rich text as one entry in a dictionary and it's not acceptable  to change the stored text.

To solve this problem, rich text is mostly a facade that changes immutable text when user calls a mutating method.

Property setter IXLRichText.Text didn't call "update-immutable-rich-text" method and thus text wasn't changed.

Fixes #2430