Fix evaluation of LOG10 function
The parser doesn't know anything about functions, but grammar does have
cell functions, so it is better to keep them and use them directly as
function names. That way, LOG10 will be evaluated and other functions
will evalaute to #REF!.

The FormulaFlags were deleted, because there is no point in storing them.
The plan is to parse formula on each evaluation instead of constructing
AST (current approach is a legacy from times of XLParser). Fixes #2229.