Remove expression cache
The cache doesn't really serve any purpose. It used a weak reference, but
other places don't reference it.

Plus the plan is to parse formula each time it is necessary. Formulas are
rarely reused, because that requires same text representation. That might
be relatively frequest for R1C1, but not for A1 formulas.

The common ClosedXML pattern is to load, do something, calculate maybe
once, save. There isn't much benefit in reusing formula ast trees.