Make getting uint from XML is error suppressable
We can get incorrect values for uint attributes (e.g. "-4") and
generraly speaking, Excel supresses error values and just treats them as
missing.

We do want to copy that behavior if possible (though it might be prudent to
write them to log once logging is available). The reading code through
XmlTreeReader.GetUInt() extension method will still throw error on missing
attribute, but not if there is a default value.