Switch to codegen pivot cache records reader
This new reader doesn't load records to memory, but is fast-forward. That
means DOM of the cache records is not loaded to memory, which saves memory
and should be faster.

The generated code in .g.cs is outputted by CodeGen project. It will be at
later date generated automatically by a source generator.