Add Context to XmlTreeParser
The context is a common pattern in event based parsers that enables
selection of correct path, when type is used in as child of different types.

Example: CT_Xf <xt> element is used in both <cellStyleXfs> and <cellXfs>.
When the hook is called, it must know whether it is in <cellStyleXfs> (named
styles) or <cellXfs> (formatting records).