Separate out a mutable part of cell format
The cell style is supposed to be mutable, but immutable cell format
readonly record can't have mutable class inside it. Therfore we use
a key to connect them through external dictionary.

Because format will be modified throguh API object, it's not really
a problem. The dictionary will be suppplied through parameter where
necessary.