Add a RefMod for shifting references on delete
The final goal is to replace IXLCell regex logic for shifting
references during structural changes. There are four changes:
* Insert, shift right
* Insert, shift down
* Delete, shift up
* Delete, shift left

This is only the first change that deals with the delete+shift up. It
should be accurate to Excel and it handles even deletion of area, not
just deletion of rows.