Remove unnecessary string allocations
The original code always allocated a new string on a heap, even though
it wasn't really necessary. If there is no change in position or size
due to deletion of an area, we can just return symbol created from the
original formula text. Because formula is already existing alllocated
string, it saves quite a few of allocations.