Refactor shift references on delete refmod
The new version is similar to the insert version. The previous version
that used TrySubtract didn't properly account for a case where there
was a area deleted through the middle of reference. In that case there
were 2 conincident lines, but thanks to shift up/left on delete,
the gap was removed.

This is a very common occurance that happens when whole row or column
is deleted.