RefersTo in a defined name can't be empty
This is a followup for #2484. The root cause there was because invalid
reference (Sheet1!#REF!) was stripped during shifting and the result
was an empty string. The FormulaParser in the RefersTo then threw
an exception.

Instead of throwing hard-to-reason ParsingException 'Error at char 0 of
'': Unexpected token EofSymbolId.', make exception easier to understand.