Add skeleton of Style part writer
This is the a skeleton writer, sufficient to write a blank file. It
doesn't yet have many styles elements and required features.

The code itself is not used anywhere, the XLWorkbook_Save.cs has to be
modified to use it.

The writer takes the all elements from XLWorkbookStyles repository,
skips the unused items and "compresses" them to a map of sequential
ids. The idea is to keep the original order in which the items were
read from the original file.