Add a constant to switch between style writers
The old writer is going the way of a dodo. So do internal styles
internal structures. But in the meantime, we need to be able to
switch from old to new.

New writer is not ready, but we need a way to switch from old
writer to new one, we can improve new one without deleting old
one.

The STYLES_REWORK is the constant that will be used to switch.