Add workbook reference to columns
We need access to cells and styles. Both are workbook. Other ways are
problematic (e.g. what if XLColumns is a set of no columns) and require
extra paths.

The longer term goal is to provide workbook to the XLCells. XLCells now
need to instantiate XLCell for each cells and it puts a heavy pressure
on GC.