Add workbook reference to range API objects
This change ensures range API objects have access to workbook.  There
are two key reasons:

Styles - We need access to XLWorkbookStyles to modify a style (e.g.,
through IXLCells.Style). That is a property on a XLWorkbook. So
basically a prerequisite for styles rework.

Performance - basically all ranges are using XLCell to modify
the values of a cell. That was fine when it was actual object that
contained all the data, but now XLCell is basically a proxy into
a sparse array of slices. With access to workbook, it will be possible
to skip XLCell proxy creation and work directly with cell
collection/slices. That means possibility of bulk operations and
significantly less pressure on GC.

Additionally, it makes life easier. Cell range lists (XLCells,
XLColumns, XLRows, XLRanges...) can be empty under certain
circumstances. In such cases, there is usually no connection to rest of
workbook. E.g. if user tries to add a range to empty range list, there
would be no way to check if it is from same workbook (not that we do
that, but we should).

The IXLStylized interface had its RangesUsed property type changed to
IEnumerable<IXLRange>. The origina IXLRange required a workbook and
that was problematic in XLStylizedEmpty class. Conversion doesn't
change functionality (used only for setting border), plus it's internal
interface.