Format slice can modify format
It is now possible to modify format of cells in a format slice through
a method that modified format of cells in an area.

This is a skeleton of modification with heavy limitations (only through
XLCell, only font, no containers, no cross determination, no cascade
resolver).

But it does allow to read a workbook with a non-default formatted cell,
modify it and save it.