Add XLColumnArea
The column area is a structure representing a column. It will be used
in various places where I need a column  alongside with an area.
The area doesn't move and therefore column shouldn't either. But
XLColumn does move. So I need a structure that represents just
a column that doesn't move.

I could use XLBookArea, but it seems tacky. Row numbers can be anything
and it just doesn't feel right.