Add (reworked) formatting to a workbook
It is now possible to use API object to modify formatting of a workbook
and have it propagate below to all other formatting containers (sheets,
columns, rows, cells).

The cell format has been switched to not-nullable, because it better
represents the actual functional structure used by formatting. Each
formatting record is completely specified.

The XLFontKey also has two incorrect keys. The charset was fixed.
XLFontCharSet.Ansi is the default/zero value, not
the XLFontCharSet.Default). The color hasn't been fixed and is left for
when style infrastructure will be switched.