Replace InnerStyle with StyleValue
Setting the InnerStyle for these occurances (XLCell, XLColumn, XLRow,
XLWorksheet) only calls XLStylizedbase.SetStyle without the propagate
flag -> ultimately it sets StyleValue by itself.

This is a step in style rework, because it is a useless preperty that
can be removed and make it simpler to switch to reworked version.