Add font properties to reworked style
Keep interface implementation in a separate file, so it's clear what is
used internaly and what not. Many classes have massive regions for
that, but I don't find it very readable plus stylecop recommends to do
it that way.

The other thing is the explicit interface implementation. The internal
properties are internal, so they can't implement interface and it makes
the separation very clear. The explicit implementation also allows to
return different types even fro props with same name.

We will see how it goes, but I find the region sections rather
confusing, mixing internal and external API.