Style rework: Fix rows inside borders
The inside borders in Excel (and in general) means set border
to everything but lines that touch outside border.

When IXLStyle was setting inside border to rows, it only set vertical
(left and right) borders. That works when there is only one row, but
not when there are two or more.

This change fixes the problem as part of style-rework. It basically
detects rowspans and sets top row, center rows and bottom row. There
is a special path for one-row, because it doesn't fit into top/center/
bottom border changes.

There is also a small change in a saving code to detect that row doesn't
have default style, but that will be refactored later. It is there only so it's
possible to verify borders were correctly created.