CF: Deduplicate code
CF has an optional dxfId attribute. DxfId is not necessary for all
types of CF (color scale, data bar and icon set).

The logic in saving converters had duplicated logic to set
the dxfId for CF types that do need it. This change moves the dxfId
determination logic to a one place, so it is not duplicated across
many converters.

Having it in a single place makes it easier to prepare style rework
conditional compiling, than to do the same thing in 12 places.