Data Validations: Replace IXRange in internal structures
This is a part of an ongoing initiative to make various objects more
self-contained and modular.

The original code was using IXLRange to determine which cells should
be affected by data validations. That has several problems:

* IXLRange magic shifting infrastructure doesn't work areas, only on
  rows/columns
* There are various subtle bugs
* XLDataValidation logic was leaking to XLWorksheet. DV should know
  best how to react to various structural changes.
* IXLRange doesn't fit into alwasy-valid model

Thus the internal property that determines which cells are affected was
switched to immutable XLAreaList (a sequence of areas) and DVs now
implement ISheetLister to properly adjust its internal structure
(shift/shrink/extend cells affected by DV, possibly delete) when there
is a structural change.

Fixes issue #2794