Enforce one-dv-per-cell invariant
When the the XLDataValidation had the IXLRanges replaced with
the XLAreaList, the DataValications also stopped enforcing
one-dv-per-cell invariant. There were also no failing tests for
the invariant.

The switch and removal was done to fix data validation shifting,
but invariant wasn't implemented in the change to keep scope
reasonable.

To make sure the invariant is enforced, adding an area to data
validation is done in a single function in XLDataValidations.
The collection contains all DVs in a sheet and adjusts/removes
DVs that no longer have any area.

The invariant has to be enforced only when area is added to DV,
removal of areas can't violate invariant. Thus IXLDV.ClearRanges()
and other area removals can be done directly in XLDataValidation.

This is basically emergecy repair, in the long run, whole conceptual
model of DVs + API should be replaced with something less concerned
with actual data structures and more with behavior. Addition or removal
of XLDataValidations should be basically transparent to user.

Example:
  ws.Range("A1:C3").DataValidations.List("one,two")
  ws.Range("A1:C3").DataValidations.Clear()

And all this DV uses these areas should be hidden away. I think
providing API that treats XML DV element as an API object is a mistake.