Hyperlinks: Allow multiple links in same area
Original required one area to have only one links. That threw
an exception during structural changes, when one link was shifted to
not-yet shifted link.

Internal structure was modified. It is now held in:
* a simple list (roughly equivalent of elements in a worksheet part)
* index: link -> area (used to find are affected by a link)
* index: area -> links (used to find links for cell/ranges).

The index:area->links uses a R-tree. This was selcted because it
provides a good performance for various use cases and I already depend
on a library that implements it.

Overall, the changes are directed toward a state where hyperlinks can
be for areas instead of cells and future API rework.

Resolves #2801